ARMOA018-446

Agregar decimales a la cantidad en todo los lugares donde aparezca

+++ SustitucionMOAModel/Dto/Compras/SolpDto.cs
@@ -357,7 +357,7 @@ namespace SustitucionMOAModel.Dto
         public TablaSapDto Unidad { get; set; }
         public string UnidadComprasDescripcion { get; set; }
         public string UnidadDescripcion { get; set; }
-        public int? CantidadCotizacion { get; set; }
+        public decimal? CantidadCotizacion { get; set; }
         public string UnidadCotizacionDescripcion { get; set; }
         public int? UnidadCotizacionId { get; set; }
         public string MonedaCotizacionDescripcion { get; set; }
