+++ SustitucionMOAModel/Entities/CotizacionSubPosicion.cs
@@ -14,7 +14,7 @@ namespace SustitucionMOAModel.Entities
         public int Id { get; set; }
         public int CotizacionPosicion_Id { get; set; }
         public int SolpSubPosicion_Id { get; set; }
-        public int? Cantidad { get; set; }
+        public decimal? Cantidad { get; set; }
         public int? UnidadDeMedida_Id { get; set; }
         public int? Moneda_Id { get; set; }     
         public decimal? Precio { get; set; }
